+++ packages/cf-deploy-config-writer/src/cf-writer/index.ts
@@ -1,2 +1,3 @@
 export { generateAppConfig } from './app-config';
 export { generateBaseConfig } from './base-config';
+export { generateCAPConfig } from './cap-config';